+++ 0015-js-structure/js-structure_quiz-app/components/CardList/CardList.js
@@ -1,25 +1,15 @@
-export const cards = [
-  {
-    question:
-      "In the Kingdom Hearts series who provides the english voice for Master Eraqus?",
-    answer: "Mark Hamill",
-    tags: ["kingdomheart", "mastereraqus", "english voices"],
-    isBookmarked: false,
-  },
-  {
-    question:
-      "Daniel Radcliffe became a global star in the film industry due to his performance in which film franchise?",
-    answer: "Harry Potter",
-    tags: ["hewhomustnotbenamed", "danielradcliffe", "abracadabra"],
-    isBookmarked: false,
-  },
-  {
-    question:
-      'In the game series "The Legend of Zelda" what is the name of the main character?',
-    answer: "Link",
-    tags: ["linktothepast", "linkbetweenworlds", "linkbetweenages"],
-    isBookmarked: false,
-  },
-];
 
-export default cards;
+import cards from '../CardList/CardArray.js'
+import Card from '../Card/Card.js'
+
+export default function CardList() {
+  const cardList = document.createElement("section");
+  cardList.classList.add("card-list");
+
+  cards.forEach((card) => {
+    const cardElement = Card(card);
+    cardList.append(cardElement);
+  });
+
+  return cardList;
+}
